+++ packages/plugin-headings/web/src/icons/index.ts
@@ -1,0 +1,3 @@
+export { default as TitleIcon } from './TitleIcon';
+export { default as TitleOneIcon } from './TitleOneIcon';
+export { default as TitleTwoIcon } from './TitleTwoIcon';